web/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
WTForms
pyjwt
PyMySQL
gunicorn